+++ lib/spree_multi_domain/factories.rb
