+++ spec/spec_helper.rb
@@ -14,7 +14,8 @@ Dir[File.join(File.dirname(__FILE__), "support/**/*.rb")].each {|f| require f }
 require 'spree/testing_support/factories'
 require 'spree/testing_support/controller_requests'
 require 'spree/testing_support/authorization_helpers'
-require 'factories'
+
+require 'spree_multi_domain/factories'
 
 RSpec.configure do |config|
   # == Mock Framework
